requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asyncvnc==1.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
discord.py @ git+https://github.com/Rapptz/discord.py@a5d03d4a1e6f4314518f6f55d8b7130ab3f4336e
et-xmlfile==1.1.0
Flask==3.0.2
Flask-Cors==4.0.0
frozenlist==1.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
keysymdef==1.2.0
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
pdfrw==0.4
pillow==10.4.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
reportlab==4.2.2
requests==2.31.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
